+++ src/Test/Aeson/Internal/GoldenSpecs.hs
@@ -23,6 +23,7 @@ import           Control.Monad
 import           Data.Aeson
 import           Data.Aeson.Encode.Pretty
 import           Data.ByteString.Lazy hiding (putStrLn)
+import           Data.Int (Int32)
 import           Data.Proxy
 import           Data.Typeable
 
@@ -170,9 +171,9 @@ mkFaultyReencodedFile (TypeNameInfo {typeNameTypeName,typeNameModuleName, typeNa
 -- | Create a number of arbitrary instances of a type
 -- a sample size and a random seed.
 mkRandomSamples :: forall a . Arbitrary a =>
-  Int -> Proxy a -> Int -> IO (RandomSamples a)
+  Int -> Proxy a -> Int32 -> IO (RandomSamples a)
 mkRandomSamples sampleSize Proxy rSeed = RandomSamples rSeed <$> generate gen
   where
     correctedSampleSize = if sampleSize <= 0 then 1 else sampleSize
     gen :: Gen [a]
-    gen = setSeed rSeed $ replicateM correctedSampleSize (arbitrary :: Gen a)
+    gen = setSeed (fromIntegral rSeed) $ replicateM correctedSampleSize (arbitrary :: Gen a)
